+++ src/lib/components/Homepage/ConferenceDetails.svelte
